+++ internal/serve/sep24frontend/app/src/helpers/localStorageSavedLanguage.ts
@@ -1,0 +1,10 @@
+import { LOCAL_STORAGE_SAVED_LANGUAGE } from "@/config/settings";
+
+export const localStorageSavedLanguage = {
+  get: () => {
+    return localStorage.getItem(LOCAL_STORAGE_SAVED_LANGUAGE);
+  },
+  set: (language: string) => {
+    return localStorage.setItem(LOCAL_STORAGE_SAVED_LANGUAGE, language);
+  },
+};